requirements.txt
streamlit
numpy
pandas
PyPDF2
python-docx
openai
pdfplumber




